Commit all files, await feedback
A commit of all files until I have a meeting with my mentor. I have removed the underline from the hypertext in the dropdown menu to improve the look, and made some other minor changes to wording, added a shadow to the submit button to make it more obvious etc., also updated the JavaScript file so that the X and Y coordinated reflect the "clicked" position on the canvas instead of the position on the entire page. I also removed superfluous code and tidied it up a bit.